基座位置添加，炮塔的搭建、升级、拆除初步设计

+++ Classes/MAP.h
@@ -20,7 +20,7 @@ protected:
 
 	CC_SYNTHESIZE(int, currentLife, CurrentLife);        //当前生命
 
-	int MonsterNum;
+	size_t MonsterNum;
 
 	int money;              //钱数
